+++ Packages/com.mimylab.fukuroudon/Runtime/ActiveRelay/Scripts/ActiveRelayToComponent.cs
@@ -9,10 +9,9 @@ namespace MimyLab.FukuroUdon
     using UdonSharp;
     using UnityEngine;
     using UnityEngine.Animations;
-    using VRC.Dynamics;
     using VRC.SDK3.Dynamics.Constraint.Components;
-    using VRC.SDK3.Dynamics.Contact.Components;
     using VRC.SDK3.Dynamics.PhysBone.Components;
+    using VRC.SDK3.Dynamics.Contact.Components;
     using VRC.Udon;
 
     [HelpURL("https://github.com/mimyquality/FukuroUdon/wiki/Active-Relay#activerelay-to-component")]
@@ -28,26 +27,40 @@ namespace MimyLab.FukuroUdon
         [SerializeField]
         private bool _invert = false;
 
+        [SerializeField, HideInInspector]
+        private UdonSharpBehaviour[] _udonSharpBehaviours = new UdonSharpBehaviour[0];
+
 #if !COMPILER_UDONSHARP && UNITY_EDITOR
         private void OnValidate()
         {
-            var count = 0;
-            var tmp = new Component[_components.Length];
+            var tmp_Components = new Component[_components.Length];
+            var componentCount = 0;
+            // U# はプロキシー概念があるので別途 UdonSharpBehaviour 配列を用意して管理する
+            var tmp_udonSharpBehaviours = new UdonSharpBehaviour[_components.Length];
+            var usbCount = 0;
             foreach (var component in _components)
             {
                 if (ValidateComponentType(component))
                 {
-                    tmp[count++] = component;
+                    tmp_Components[componentCount++] = component;
+                    
+                    if (component is UdonSharpBehaviour usb)
+                    {
+                        tmp_udonSharpBehaviours[usbCount++] = usb;
+                    }
                 }
             }
-            System.Array.Resize(ref tmp, count);
-            _components = tmp;
+            System.Array.Resize(ref tmp_Components, componentCount);
+            _components = tmp_Components;
+            System.Array.Resize(ref tmp_udonSharpBehaviours, usbCount);
+            _udonSharpBehaviours = tmp_udonSharpBehaviours;
         }
 
         private bool ValidateComponentType(Component component)
         {
             if (!component) { return false; }
 
+            var type = component.GetType();
             if (component is Collider) { return true; }
             if (component is Renderer) { return true; }
             // 個別に羅列の必要ある
@@ -72,7 +85,7 @@ namespace MimyLab.FukuroUdon
             if (component is Camera) { return true; }
             if (component is Animator) { return true; }
             if (component is UdonBehaviour) { return true; }
-            //if (component is UdonSharpBehaviour) { return true; } // なんとかする
+            if (component is UdonSharpBehaviour) { return true; }
             if (component is VRCParentConstraint) { return true; }
             if (component is VRCPositionConstraint) { return true; }
             if (component is VRCRotationConstraint) { return true; }
@@ -155,6 +168,11 @@ namespace MimyLab.FukuroUdon
                 // else if (type.IsSubclassOf(typeof(Behaviour))) { var downCasted = (Behaviour)component; downCasted.enabled = value; }
                 // else if (type == typeof(AudioListener)) { var downCasted = (AudioListener)component; downCasted.enabled = value; }
                 // else if (type == typeof(Cloth)) { var downCasted = (Cloth)component; downCasted.enabled = value; }
+            }
+            foreach (var udonSharpBehaviour in _udonSharpBehaviours)
+            {
+                if (!udonSharpBehaviour) continue;
+                udonSharpBehaviour.enabled = value;
             }
         }
     }